update bsd and darwin tests
